requirements.txt
streamlit
instaloader
pillow
